requirements.txt
alembic==1.4.1
aniso8601==9.0.1
asgiref==3.4.1
azure-core==1.16.0
azure-storage-blob==12.8.1
boto3==1.18.11
botocore==1.21.11
click==7.1.2
cloudpickle==1.6.0
coverage==5.5
databricks-cli==0.14.3
dill==0.3.4
docker==5.0.0
dotmap==1.3.24
elastic-apm==6.3.3
elasticsearch==7.10.0
faiss-cpu==1.7.1.post2
farm==0.8.0
farm-haystack==0.9.0
fastapi==0.68.0
greenlet==1.1.0
grpcio==1.37.1
grpcio-tools==1.37.1
gunicorn==20.1.0
h11==0.12.0
httptools==0.2.0
huggingface-hub==0.0.8
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
joblib==1.0.1
langdetect==1.0.9
lxml==4.6.3
Mako==1.1.4
MarkupSafe==2.0.1
mlflow==1.13.1
mmh3==3.0.0
more-itertools==8.8.0
msrest==0.6.21
networkx==2.6.2
nltk==3.6.2
numpy==1.21.1
oauthlib==3.1.1
pandas==1.3.1
pluggy==0.13.1
prometheus-flask-exporter==0.18.2
protobuf==3.17.3
psutil==5.8.0
psycopg2-binary==2.9.1
py==1.10.0
pydantic==1.8.2
pymilvus==1.1.2
python-docx==0.8.11
python-editor==1.0.4
python-multipart==0.0.5
pytz==2021.1
querystring-parser==1.2.4
rdflib==6.0.0
regex==2021.7.6
requests-oauthlib==1.3.0
s3transfer==0.5.0
sacremoses==0.0.45
scikit-learn==0.24.2
scipy==1.7.0
sentencepiece==0.1.96
seqeval==1.2.2
sklearn==0.0
SPARQLWrapper==1.8.5
SQLAlchemy==1.4.22
SQLAlchemy-Utils==0.37.8
sqlparse==0.4.1
starlette==0.14.2
tabulate==0.8.9
threadpoolctl==2.2.0
tika==1.24
tokenizers==0.10.3
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.8.0+cpu 
tox==3.24.1
transformers==4.6.1
ujson==4.0.2
uvicorn==0.14.0
uvloop==0.14.0
validators==0.18.2
weaviate-client==3.0.0
websocket-client==1.1.0
Werkzeug==0.16.1
streamlit==0.85.1
pymupdf==1.18.16